rename shift_combinations_key.py file name
