+++ main.py
@@ -8,7 +8,7 @@ from constraints_file import generate_shift_employee_combinations, add_exactly_o
 from models.employees.employee import Employee
 from models.employees.employee_priority_enum import EmployeePriorityEnum
 from models.employees.employee_status_enum import EmployeeStatusEnum
-from models.shifts.frozen_shift_combinations_key import ShiftCombinationsKey
+from models.shifts.shift_combinations_key import ShiftCombinationsKey
 from models.shifts.shift import Shift
 from models.shifts.shifts_types_enum import ShiftTypesEnum
